refactor PartnerAdmin partner state
