+++ app/api.ts
@@ -7,11 +7,11 @@ import {
 } from '@reduxjs/toolkit/query/react';
 import { auth } from '../config/firebase';
 import { PARTNER_ACCESS_CODE_STATUS } from '../constants/enums';
+import { Partner } from '../constants/partners';
 import { delay } from '../utils/delay';
 import { Course, Courses } from './coursesSlice';
 import { PartnerAccess, PartnerAccesses } from './partnerAccessSlice';
 import { PartnerAdmin } from './partnerAdminSlice';
-import { Partner } from './partnerSlice';
 import { RootState } from './store';
 import { User } from './userSlice';
